Maintenance: create a higher order function to fetch an item
- Create makeFetchItemAction and makeFetchItemsAction and convert
fetchThreads, posts etc to use this higher order function.
- Add a new parameter 'once' to this action with the default set to
false in readiness for implementing a notification system.
- The impetus for refactoring thus is the Forum page fetches the item twice,
in the route navigation guard and in the created hook, we only want to subscribe
to this firestore snapshot once. If the once param is set to true as we do in
the navigation guard, we call the unsubscribe method before fetching the item
in the fetchItem action.