Feature: notifications
- When a post is edited or created display a 'recently updated' notification.
- Check the server timestamp of the thread before setting the notification because
if a post is added, 2 notifications are created - once for post and once for
thread. When a post is updated the post.edited field is updated with a server
timestamp which means onSnaphot first fires for the local write with the timestamp set to
null and then fires again when it is written server side using the server timestamp. So we
only create the recently updated notification if the edited.at field is not null.