Ensure the replies count makes sense.
Issue - when a user creates a thread it is counted as a reply, but
strictly speaking it is not a reply. Create a "firstInThread" property,
set to true when creating a thread otherwise defaulting to false, and
only append the user id to the thread if it is not the firstInThread.

Update the display of the count - if 1 should be reply not replies. Same
for 'contributors'.